+++ src/main/java/com/epam/reportportal/utils/properties/ListenerProperty.java
@@ -148,11 +148,17 @@ public enum ListenerProperty {
     TRUNCATE_REPLACEMENT("rp.truncation.replacement", false),
     TRUNCATE_ITEM_NAME_LIMIT("rp.truncation.item.name.limit", false),
     TRUNCATE_ATTRIBUTE_LIMIT("rp.truncation.attribute.limit", false),
+    ATTRIBUTE_NUMBER_LIMIT("rp.attribute.limit", false),
 
     /**
-     * Enable/Disable the feature to truncate Stack Traces of exceptions that being logged to ReportPortal. Default value: <code>true</code>.
+     * Toggle the feature to truncate Stack Traces of exceptions that being logged to ReportPortal. Default value: <code>true</code>.
      */
     EXCEPTION_TRUNCATE("rp.truncation.exception", false),
+
+	/**
+	 * Toggle the feature to replace basic binary characters with \uFFFD char. Default value: <code>true</code>.
+	 */
+	REPLACE_BINARY_CHARACTERS("rp.sanitization.replace.binary", false),
 
     // Issue reporting properties
     /**